+++ src/main/java/uk/nhs/prm/repo/ehrtransferservice/exceptions/HttpException.java
@@ -1,10 +1,6 @@
 package uk.nhs.prm.repo.ehrtransferservice.exceptions;
 
 public class HttpException extends Exception {
-    public HttpException(String message, Exception cause) {
-        super(message, cause);
-    }
-
     public HttpException() {
         super();
     }
