+++ src/main/java/uk/nhs/prm/repo/ehrtransferservice/exceptions/database/ConversationNotPresentException.java
@@ -1,14 +1,14 @@
-package uk.nhs.prm.repo.ehrtransferservice.exceptions;
+package uk.nhs.prm.repo.ehrtransferservice.exceptions.database;
 
 import uk.nhs.prm.repo.ehrtransferservice.exceptions.base.DatabaseException;
 
 import java.util.UUID;
 
-public class TransferRecordNotPresentException extends DatabaseException {
+public class ConversationNotPresentException extends DatabaseException {
     private static final String EXCEPTION_MESSAGE =
         "No transfer present for Inbound Conversation ID %s";
 
-    public TransferRecordNotPresentException(UUID inboundConversationId) {
+    public ConversationNotPresentException(UUID inboundConversationId) {
         super(EXCEPTION_MESSAGE.formatted(inboundConversationId.toString()));
     }
 }
