+++ src/main/java/uk/nhs/prm/repo/ehrtransferservice/exceptions/database/ConversationUpdateException.java
@@ -1,0 +1,18 @@
+package uk.nhs.prm.repo.ehrtransferservice.exceptions.database;
+
+import uk.nhs.prm.repo.ehrtransferservice.exceptions.base.DatabaseException;
+
+import java.util.UUID;
+
+public class ConversationUpdateException extends DatabaseException {
+    private static final String EXCEPTION_MESSAGE =
+        "The conversation could not be updated with Inbound Conversation ID %s";
+
+    public ConversationUpdateException(UUID inboundConversationId) {
+        super(EXCEPTION_MESSAGE.formatted(inboundConversationId));
+    }
+
+    public ConversationUpdateException(UUID inboundConversationId, Throwable throwable) {
+        super(EXCEPTION_MESSAGE.formatted(inboundConversationId), throwable);
+    }
+}
